+++ Source/InMobiAdapter.swift
@@ -111,7 +111,12 @@ final class InMobiAdapter: NSObject, PartnerAdapter {
         case .banner:
             return try InMobiAdapterBannerAd(adapter: self, request: request, delegate: delegate)
         default:
-            throw error(.loadFailureUnsupportedAdFormat)
+            // Not using the `.adaptiveBanner` case directly to maintain backward compatibility with Chartboost Mediation 4.0
+            if request.format.rawValue == "adaptive_banner" {
+                return try InMobiAdapterBannerAd(adapter: self, request: request, delegate: delegate)
+            } else {
+                throw error(.loadFailureUnsupportedAdFormat)
+            }
         }
     }