+++ Source/InMobiAdapterBannerAd.swift
@@ -30,12 +30,19 @@ final class InMobiAdapterBannerAd: InMobiAdapterAd, PartnerAd {
     /// - parameter completion: Closure to be performed once the ad has been loaded.
     func load(with viewController: UIViewController?, completion: @escaping (Result<PartnerEventDetails, Error>) -> Void) {
         log(.loadStarted)
-        
+
+        // Fail if we cannot fit a fixed size banner in the requested size.
+        guard let size = fixedBannerSize(for: request.size ?? IABStandardAdSize) else {
+            let error = error(.loadFailureInvalidBannerSize)
+            log(.loadFailed(error))
+            return completion(.failure(error))
+        }
+
         // Save completion for later
         loadCompletion = completion
         
         // Create the banner
-        let frame = CGRect(origin: .zero, size: request.size ?? IABStandardAdSize)
+        let frame = CGRect(origin: .zero, size: size)
         let ad = IMBanner(frame: frame, placementId: placementID, delegate: self)
         ad.shouldAutoRefresh(false)
         inlineView = ad
@@ -63,7 +70,14 @@ extension InMobiAdapterBannerAd: IMBannerDelegate {
     func bannerDidFinishLoading(_ banner: IMBanner) {
         // Report load success
         log(.loadSucceeded)
-        loadCompletion?(.success([:])) ?? log(.loadResultIgnored)
+
+        var partnerDetails: [String: String] = [:]
+        if let loadedSize = fixedBannerSize(for: request.size ?? IABStandardAdSize) {
+            partnerDetails["bannerWidth"] = "\(loadedSize.width)"
+            partnerDetails["bannerHeight"] = "\(loadedSize.height)"
+            partnerDetails["bannerType"] = "0" // Fixed banner
+        }
+        loadCompletion?(.success(partnerDetails)) ?? log(.loadResultIgnored)
         loadCompletion = nil
     }
     
@@ -79,5 +93,22 @@ extension InMobiAdapterBannerAd: IMBannerDelegate {
         // Report click
         log(.didClick(error: nil))
         delegate?.didClick(self, details: [:]) ?? log(.delegateUnavailable)
+    }
+}
+
+// MARK: - Helpers
+extension InMobiAdapterBannerAd {
+    private func fixedBannerSize(for requestedSize: CGSize) -> CGSize? {
+        let sizes = [IABLeaderboardAdSize, IABMediumAdSize, IABStandardAdSize]
+        // Find the largest size that can fit in the requested size.
+        for size in sizes {
+            // If height is 0, the pub has requested an ad of any height, so only the width matters.
+            if requestedSize.width >= size.width &&
+                (size.height == 0 || requestedSize.height >= size.height) {
+                return size
+            }
+        }
+        // The requested size cannot fit any fixed size banners.
+        return nil
     }
 }